+++ src/Relations/HasManyRelation.php
@@ -2,13 +2,18 @@
 
 namespace Swis\JsonApi\Client\Relations;
 
-class HasManyRelation extends AbstractManyRelation
+use Swis\JsonApi\Client\Interfaces\TypedRelationInterface;
+use Swis\JsonApi\Client\Relations\Traits\HasType;
+
+class HasManyRelation extends AbstractManyRelation implements TypedRelationInterface
 {
+    use HasType;
+
     /**
      * @param string $type
      */
     public function __construct(string $type)
     {
-        $this->type = $type;
+        $this->setType($type);
     }
 }